More styling and some refactor
