+++ index.html
@@ -14,7 +14,7 @@
     <meta http-equiv="Expires" content="0">
   </head>
   <body>
-    <div id="app" class="container py-4 px-3 mx-auto"></div>
+    <div id="app" class="container bg-light"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
